Created second branch of codebase to address synching problems. Testing synching based on frames or "cycles". This gives reliable periodic collision, which is fundamental.
Some potential problems are with importing wavs/oggs, which have a specific BPM. Cycle-based synching is not guaranteed to be accurate in terms of milliseconds (time per cycle is variable). Plan is to implement composer framework within the game, which sequences based on cycles.

Other small changes/housekeeping. E.g. don't display mouse cursor outline unless on the game grid