Significant update and changes.
Decided to move to a more abstract graphics look and feel. Initial 8-bit graphics allowed me to see that a dedicated artist would be needed, or a gazillion hours of trial and learning.

Following the influence of Midinous, I've changed the bugs on tracks theme, to a sparks on paths theme. And the modelling has been updated accordingly. Instead of 'tracks' being composed of lines of grid rects, 'paths' are now aligned to the grid lines. Added a DirectedGraph class to manage the paths, which are essentially collections of edges (node to node).

Another significant change is with the grid_offset that was being passed around everywhere. The level now has a 'view', which has it's own frame of reference - meaning it starts at (0, 0) and ends at (width, heigth). Everything gets calculated in view-coords and the view is just blitted onto the gamescreen. Waaaay simpler.

Sad to see the beatbugs go, but onwards and... mmm... outwards?

This commit also includes the change from ticks-based to frame-based synchronisation. The idea is to build up the template tunes in TC (TuneCraft - new working name). This way synchronisation between template tunes and puzzle tunes can be exact.

Added screen transitions. So, screens can fade from one to the next in a variety of ways.

Added GameScreen and MenuScreen frameworks and moved UI python modules into their own folder.